Bug: StaticCredentialsProvider,StatusCode="Unauthenticated" bug
