+++ src/Ydb.Sdk/src/Auth/ICredentialsProvider.cs
@@ -3,4 +3,6 @@
 public interface ICredentialsProvider
 {
     string? GetAuthInfo();
+    
+    void ProvideConfig(DriverConfig driverConfig) {}
 }
