--- src/Ydb.Sdk/src/Auth/IUseDriverConfig.cs
@@ -1,6 +1,0 @@
-namespace Ydb.Sdk.Auth;
-
-public interface IUseDriverConfig
-{
-    public Task ProvideConfig(DriverConfig driverConfig);
-}
