+++ src/Ydb.Sdk/src/DriverConfig.cs
@@ -31,6 +31,8 @@ public class DriverConfig
         DefaultTransportTimeout = defaultTransportTimeout ?? TimeSpan.FromMinutes(1);
         DefaultStreamingTransportTimeout = defaultStreamingTransportTimeout ?? TimeSpan.FromMinutes(10);
         CustomServerCertificate = customServerCertificate;
+
+        Credentials.ProvideConfig(this);
     }
 
     private static string FormatEndpoint(string endpoint)
